Retina kernel treats time and dt as double-precision.  Retina parameters beginStim and endStim are now doubles as well.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@6410 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/kernels/Retina_params.h
@@ -4,8 +4,8 @@
 typedef struct Retina_params_ {
    float probStim;
    float probBase;
-   float beginStim;
-   float endStim;
+   double beginStim;
+   double endStim;
    float burstFreq;          // frequency of bursts
    float burstDuration;      // duration of each burst, <=0 -> sinusoidal
